requirements.txt
# 基础依赖
python-dotenv>=0.19.0
requests>=2.28.0
tqdm>=4.64.0

# 视频处理
moviepy>=1.0.3

# AI接口
openai>=1.0.0
httpx>=0.24.0
websocket-client>=1.5.0
pandas>=1.5.0

# UI界面
PySide6>=6.4.0